refactor: Renamed TrackPlotTab properties, Lineage -> LineageGraph
