+++ src/components/Tabs/Plots/PlotsTab.tsx
@@ -5,7 +5,7 @@ import type { ShowAlertBannerCallback } from "src/components/Banner";
 import SelectionDropdown from "src/components/Dropdowns/SelectionDropdown";
 import type { SelectItem } from "src/components/Dropdowns/types";
 import CorrelationPlotTab from "src/components/Tabs/CorrelationPlot/CorrelationPlotTab";
-import LineageTab from "src/components/Tabs/Plots/Lineage/LineageTab";
+import LineageGraphTab from "src/components/Tabs/Plots/LineageGraph/LineageGraphTab";
 import Plot3dTab from "src/components/Tabs/Plots/Plot3d/Plot3dTab";
 import ScatterPlotTab from "src/components/Tabs/Plots/ScatterPlot/ScatterPlotTab";
 import TrackPlotTab from "src/components/Tabs/Plots/TrackPlot/PlotTab";
@@ -126,7 +126,7 @@ export default function PlotsTab(props: PlotsTabProps): ReactElement {
       </PlotContainer>
 
       <PlotContainer visible={plotTab === PlotTabType.LINEAGE}>
-        <LineageTab {...sharedProps} />
+        <LineageGraphTab {...sharedProps} />
       </PlotContainer>
     </div>
   );
