+++ src/components/Tabs/Plots/TrackPlot/TrackPlotTab.tsx
@@ -8,7 +8,7 @@ import IconButton from "src/components/Buttons/IconButton";
 import ColorRampDropdown from "src/components/Dropdowns/ColorRampDropdown";
 import LoadingSpinner from "src/components/LoadingSpinner";
 import PlotsTabToolbar from "src/components/Tabs/Plots/PlotsTabToolbar";
-import TrackPlotWrapper from "src/components/Tabs/Plots/TrackPlot/PlotWrapper";
+import TrackPlotWrapper from "src/components/Tabs/Plots/TrackPlot/TrackPlotWrapper";
 import type { SharedPlotTabProps } from "src/components/Tabs/Plots/types";
 import { useViewerStateStore } from "src/state";
 import { FlexRowAlignCenter, NoSpinnerContainer } from "src/styles/utils";
